Add diacritics-related preferences and tests.
The user can now choose wether to receive comma-below or cedilla-below diacritics and a-circumflex or i-circumflex diacritics. The default is what we have in the csv file (comma-below all-around, mostlyi-circumflex).

This change was forced by the changes operated in SIRUTA upstream. Apparently the old, i-circumflex names were never changed for most entities, thus many, but not all names with Â were converted to Î. We strongly recommend users to set_diacritics_params before starting to use the functions.

Conflicts:
	sirutalib.py
	testsiruta.py